+++ src/main/java/GDSCKNU/CitySavior/controller/ReportsController.java
@@ -1,11 +1,11 @@
 package GDSCKNU.CitySavior.controller;
 
 import GDSCKNU.CitySavior.annotation.HasFile;
-import GDSCKNU.CitySavior.dto.request.CreateReportCommentRequest;
-import GDSCKNU.CitySavior.dto.response.MapReportsResponse;
-import GDSCKNU.CitySavior.dto.response.ReportDetailResponse;
-import GDSCKNU.CitySavior.dto.request.ReportRequest;
-import GDSCKNU.CitySavior.dto.response.StatisticsResponse;
+import GDSCKNU.CitySavior.dto.reportComment.request.CreateReportCommentRequest;
+import GDSCKNU.CitySavior.dto.report.response.MapReportsResponse;
+import GDSCKNU.CitySavior.dto.report.response.ReportDetailResponse;
+import GDSCKNU.CitySavior.dto.report.request.ReportRequest;
+import GDSCKNU.CitySavior.dto.report.response.StatisticsResponse;
 import GDSCKNU.CitySavior.service.AIService;
 import GDSCKNU.CitySavior.service.ReportService;
 import GDSCKNU.CitySavior.service.StorageService;
