+++ src/main/java/GDSCKNU/CitySavior/dto/report/response/StatisticsResponse.java
@@ -1,4 +1,4 @@
-package GDSCKNU.CitySavior.dto.response;
+package GDSCKNU.CitySavior.dto.report.response;
 
 import GDSCKNU.CitySavior.domain.Category;
 import GDSCKNU.CitySavior.entity.Report;
