+++ src/main/java/GDSCKNU/CitySavior/dto/reportComment/request/CreateReportCommentRequest.java
@@ -1,4 +1,4 @@
-package GDSCKNU.CitySavior.dto.request;
+package GDSCKNU.CitySavior.dto.reportComment.request;
 
 public record CreateReportCommentRequest(
         String comment
