+++ src/main/java/GDSCKNU/CitySavior/service/ReportService.java
@@ -1,9 +1,9 @@
 package GDSCKNU.CitySavior.service;
 
-import GDSCKNU.CitySavior.dto.response.MapReportsResponse;
-import GDSCKNU.CitySavior.dto.response.ReportDetailResponse;
-import GDSCKNU.CitySavior.dto.request.ReportRequest;
-import GDSCKNU.CitySavior.dto.response.StatisticsResponse;
+import GDSCKNU.CitySavior.dto.report.response.MapReportsResponse;
+import GDSCKNU.CitySavior.dto.report.response.ReportDetailResponse;
+import GDSCKNU.CitySavior.dto.report.request.ReportRequest;
+import GDSCKNU.CitySavior.dto.report.response.StatisticsResponse;
 import java.util.List;
 import java.util.Map;
