+++ src/test/java/GDSCKNU/CitySavior/controller/ReportsControllerTest.java
@@ -4,7 +4,7 @@ package GDSCKNU.CitySavior.controller;
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.multipart;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 
-import GDSCKNU.CitySavior.dto.request.ReportRequest;
+import GDSCKNU.CitySavior.dto.report.request.ReportRequest;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.FileInputStream;
 import org.junit.jupiter.api.DisplayName;
